Split contents of app_store_connect_api.py into separate files
